Video + edits
Added green bar at top as well as scaling video.  Rest of site needs to
be switched into same scalable format.  Comapare it to riskeverything.us
that is what I am going for with the video scalability.  When ours snaps
to mobile itll obviously need to work differently but yeh